Fixed a PHP error that could occur when using the `--continue-on-error` flag on the `feed-me/feeds/queue` CLI command.

+++ src/services/Process.php
@@ -524,22 +524,24 @@ class Process extends Component
             return;
         }
 
-        $elementsToDeleteDisable = array_diff($settings['existingElements'], $processedElementIds);
+        if ($processedElementIds) {
+            $elementsToDeleteDisable = array_diff($settings['existingElements'], $processedElementIds);
 
-        if ($elementsToDeleteDisable) {
-            if (DuplicateHelper::isDisable($feed)) {
-                $this->_service->disable($elementsToDeleteDisable);
-                $message = 'The following elements have been disabled: ' . json_encode($elementsToDeleteDisable) . '.';
-            } elseif (DuplicateHelper::isDisableForSite($feed)) {
-                $this->_service->disableForSite($elementsToDeleteDisable);
-                $message = 'The following elements have been disabled for the target site: ' . json_encode($elementsToDeleteDisable) . '.';
-            } else {
-                $this->_service->delete($elementsToDeleteDisable);
-                $message = 'The following elements have been deleted: ' . json_encode($elementsToDeleteDisable) . '.';
+            if ($elementsToDeleteDisable) {
+                if (DuplicateHelper::isDisable($feed)) {
+                    $this->_service->disable($elementsToDeleteDisable);
+                    $message = 'The following elements have been disabled: ' . json_encode($elementsToDeleteDisable) . '.';
+                } elseif (DuplicateHelper::isDisableForSite($feed)) {
+                    $this->_service->disableForSite($elementsToDeleteDisable);
+                    $message = 'The following elements have been disabled for the target site: ' . json_encode($elementsToDeleteDisable) . '.';
+                } else {
+                    $this->_service->delete($elementsToDeleteDisable);
+                    $message = 'The following elements have been deleted: ' . json_encode($elementsToDeleteDisable) . '.';
+                }
+
+                Plugin::info($message);
+                Plugin::debug($message);
             }
-
-            Plugin::info($message);
-            Plugin::debug($message);
         }
 
         // Log the total time taken to process the feed
@@ -548,7 +550,7 @@ class Process extends Component
 
         Plugin::$stepKey = null;
 
-        $message = 'Processing ' . count($processedElementIds) . ' elements finished in ' . $execution_time . 's';
+        $message = 'Processing ' . ($processedElementIds ? count($processedElementIds) : 0) . ' elements finished in ' . $execution_time . 's';
         Plugin::info($message);
         Plugin::debug($message);
